requirements.txt
pyserial==3.5
PySide6==6.2.2.1
shiboken6==6.2.2.1
